refactor: AdminPage 의 상품 관리와 쿠폰 관리 UI 를 별도 컴포넌트로 분리한다
